server/requirements.txt
langchain>=0.3.0
langchain-openai>=0.3.0
langgraph>=0.2.0
langchain-core>=0.3.0
openai>=1.50.0
pydantic>=2.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
websockets>=13.0
httpx>=0.27.0
python-dotenv>=1.0.0
google-api-python-client>=2.140.0
google-auth>=2.30.0
